refactor(leaderboard): Split the leaderboard page into individual and team pages and attempt to apply ISR
